pyproject.toml
[project]
name = "advance-rag"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sahilsingh1123", email = "sahilsingh1123@gmail.com" }
]
dependencies = [
    "fastapi>=0.128.3",
    "uvicorn>=0.40.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "pgvector>=0.4.2",
    "redis>=7.1.0",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "pandas>=3.0.0",
    "celery>=5.6.2",
    "prometheus-client>=0.24.1",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "typer>=0.21.1",
    "sentence-transformers>=5.2.2",
    "scikit-learn>=1.8.0",
    "markdown>=3.10.1",
    "beautifulsoup4>=4.14.3",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "pyjwt>=2.11.0",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "sqlalchemy>=2.0.46",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-propagator-b3>=1.39.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
advance-rag = "advance_rag.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "black>=26.1.0",
    "ruff>=0.15.0",
    "mypy>=1.19.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/advance_rag"]
